:lipstick: (llm) update post purchase screen
