+++ .changeset/three-cameras-hide.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+Update post purchase success screen
